Move SnackbarSeverity to src/constants.ts
