+++ src/components/Auth.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import { Redirect } from 'react-router-dom';
-import { SnackbarSeverity } from '../App';
+import { SnackbarSeverity } from '../constants';
 
 type Props = {
   isLoggedIn: null | boolean;
