+++ src/constants.ts
@@ -1,0 +1,6 @@
+export enum SnackbarSeverity {
+    ERROR = 'error',
+    WARNING = 'warning',
+    INFO = 'info',
+    SUCCESS = 'success',
+}
